refactor : cubeState all settle down
